[CUMULUS-2471] Add /s3CredentialsReadme to distribution API (#2224)
