+++ package.json
@@ -112,6 +112,7 @@
     "file-loader": "^3.0.1",
     "got": "^11.7.0",
     "hasha": "v5.2.0",
+    "html-loader": "^1.3.2",
     "inject-loader": "^4.0.1",
     "is-ip": "^3.1.0",
     "jasmine": "^3.1.0",
