+++ packages/api/webpack.config.js
@@ -100,6 +100,10 @@ module.exports = {
           },
         ],
       },
+      {
+        test: /\.html$/i,
+        loader: 'html-loader',
+      },
     ],
   },
   target: 'node',
