+++ packages/s3-credentials-endpoint/index.js
@@ -6,17 +6,12 @@ const boom = require('express-boom');
 const cookieParser = require('cookie-parser');
 const cors = require('cors');
 const distributionRouter = require('express-promise-router')();
-const template = require('lodash/template');
-const { promisify } = require('util');
-const fs = require('fs');
-const readFile = promisify(fs.readFile);
 const {
   EarthdataLoginClient,
   EarthdataLoginError,
 } = require('@cumulus/oauth-client');
 const express = require('express');
 const hsts = require('hsts');
-const { join: pathjoin } = require('path');
 const Logger = require('@cumulus/logger');
 const morgan = require('morgan');
 const urljoin = require('url-join');
@@ -27,6 +22,7 @@ const { isAccessTokenExpired } = require('@cumulus/api/lib/token');
 const { getUserAccessibleBuckets } = require('@cumulus/cmrjs');
 const awsServices = require('@cumulus/aws-client/services');
 const { RecordDoesNotExist } = require('@cumulus/errors');
+const displayS3CredentialInstructions = require('@cumulus/api/endpoints/s3credentials-readme');
 
 const log = new Logger({ sender: 's3credentials' });
 
@@ -80,19 +76,6 @@ async function requestTemporaryCredentialsFromNgap({
     FunctionName: lambdaFunctionName,
     Payload,
   }).promise();
-}
-
-/**
- * Sends a sample webpage describing how to use s3Credentials endpoint
- *
- * @param {Object} _req - express request object (unused)
- * @param {Object} res - express response object
- * @returns {Object} express repose object of the s3Credentials directions.
- */
-async function displayS3CredentialInstructions(_req, res) {
-  const instructionTemplate = await readFile(pathjoin(__dirname, 'instructions', 'index.html'), 'utf-8');
-  const compiled = template(instructionTemplate);
-  res.send(compiled(process.env));
 }
 
 /**
