+++ packages/s3-credentials-endpoint/package.json
@@ -7,7 +7,7 @@
     "node": ">=12.18.0"
   },
   "scripts": {
-    "build": "../../node_modules/.bin/webpack && cp -Rf instructions dist/instructions && (cd dist && rm -f lambda.zip && node ../../../bin/zip.js lambda.zip index.js instructions)",
+    "build": "../../node_modules/.bin/webpack && (cd dist && rm -f lambda.zip && node ../../../bin/zip.js lambda.zip index.js)",
     "package": "npm run build",
     "test": "../../node_modules/.bin/ava",
     "test:coverage": "../../node_modules/.bin/nyc npm test"
