+++ packages/s3-credentials-endpoint/tests/test-index.js
@@ -37,7 +37,6 @@ const {
 } = require('..');
 
 const index = rewire('../index.js');
-const displayS3CredentialInstructions = index.__get__('displayS3CredentialInstructions');
 const parseBucketKey = index.__get__('parseBucketKey');
 const formatAllowedBucketKeys = index.__get__('formatAllowedBucketKeys');
 const fetchPolicyForUser = index.__get__('fetchPolicyForUser');
@@ -333,15 +332,6 @@ test('s3credentials() with a username and a client name sends the correct reques
   await s3credentials(req, res);
 
   t.is(lambdaInvocationCount, 1);
-});
-
-test('displayS3Credentials fills template with correct distribution endpoint.', async (t) => {
-  const send = sinon.spy();
-  const res = { send };
-  const expectedLink = `<a href="${process.env.DISTRIBUTION_ENDPOINT}s3credentials" target="_blank">${process.env.DISTRIBUTION_ENDPOINT}s3credentials</a>`;
-
-  await displayS3CredentialInstructions(undefined, res);
-  t.true(send.calledWithMatch(expectedLink));
 });
 
 test.serial('An s3credential request with DISABLE_S3_CREDENTIALS set to true results in a 503 error', async (t) => {
